Completed connection manager

+++ chapter_6/api/connection_manager_api.q
@@ -15,6 +15,12 @@
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++//
 
 /
+* @brief Name of this process used for communication. This value is initialized at the registration
+*  to the connection manager.
+\
+MY_ACCOUNT_NAME: `;
+
+/
 * @brief Socket used to communicate with a local connection manager defined in `connection_manager.config`.
 \
 CONNECTION_MANAGER_SOCKET: {[]
@@ -41,6 +47,13 @@ CONNECTION: flip `host`port`socket!"**i"$\:();
 \
 CONSUMER_FILTERS: 2!flip `channel`topic`sockets!"ss*"$\:();
 
+/
+* @brief Map from a target name to a unique channel.
+* @key symbol: Target user name.
+* @value symbol: Channel name used for the chat with the target.
+\
+PRIVATE_MESSAGE_CHANNEL: (`$())!`$();
+
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++//
 //                   Private Functions                   //
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++//
@@ -66,7 +79,10 @@ register_counter_party:{[host; port]
 add_consumer_filter:{[channel;topics;remote;socket]
   filters: channel,/: topics,\: $[remote; .z.w; socket];
   .log.info["add a consumer to filters"; filters];
-  `CONSUMER_FILTERS upsert/: filters;
+  {[channel;topic;socket]
+    // Replace the value (`sockets![list of int]) with the new value of the same type.
+    CONSUMER_FILTERS[(channel; topic)]:@[CONSUMER_FILTERS[(channel; topic)]; `sockets; ,; socket];
+  } ./: filters;
  };
 
 /
@@ -127,7 +143,7 @@ delete_socket_from_filters:{[socket_]
       // Last socket. Delete the filter.
       delete from `CONSUMER_FILTERS where channel = channel_, topic = topic_;
       // Other consumers are subscribing
-      CONSUMER_FILTERS[(channel_; topic_)]: CONSUMER_FILTERS[(channel_; topic_)] except socket_
+      CONSUMER_FILTERS[(channel_; topic_)]: @[CONSUMER_FILTERS[(channel_; topic_)]; `sockets; except; socket_]
     ];
   }[socket_] ./: keys_and_count; 
  };
@@ -138,10 +154,13 @@ delete_socket_from_filters:{[socket_]
 \
 .z.pc:{[socket_]
   handle: hsym `$":" sv first each exec (host; port) from CONNECTION where socket = socket_;
-  .log.info["connection dropped"; handle];
-  delete from `CONNECTION where socket = socket_;
-  // Delete the socket from filters
-  delete_socket_from_filters[socket_];
+  if[not handle ~ `:;
+    // Registered socket.
+    .log.info["connection dropped"; handle];
+    delete from `CONNECTION where socket = socket_;
+    // Delete the socket from filters
+    delete_socket_from_filters[socket_]
+  ];
  };
 
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++//
@@ -150,8 +169,12 @@ delete_socket_from_filters:{[socket_]
 
 /
 * @brief Register channel as a producer. If there are matched consumers, start communication with them.
+* @param name {symbol}: User name to connect to the connection manager.
+* @param channel {symbol}: Channel to which publish a message.
 \
 .cmng_api.register_as_producer: {[name;channel]
+  // Initialize account name if not defined
+  if[MY_ACCOUNT_NAME ~ `; MY_ACCOUNT_NAME::name];
   matched: CONNECTION_MANAGER_SOCKET (`.cmng.register; name; 1b; channel; (::));
   $[count matched;
     [
@@ -168,9 +191,14 @@ delete_socket_from_filters:{[socket_]
  };
 
 /
-* @brief Register channel as a producer. If there are matched consumers, start communication with them.
+* @brief Register channel as a consumer. If there are matched producers, start communication with them.
+* @param name {symbol}: User name to connect to the connection manager.
+* @param channel {symbol}: Channel to subscribe.
+* @topics {list of symbol}: Topics to subscribe.
 \
 .cmng_api.register_as_consumer: {[name;channel;topics]
+  // Initialize account name if not defined
+  if[MY_ACCOUNT_NAME ~ `; MY_ACCOUNT_NAME::name];
   matched: CONNECTION_MANAGER_SOCKET (`.cmng.register; name; 0b; channel; topics);
   $[count matched;
     [
@@ -184,4 +212,67 @@ delete_socket_from_filters:{[socket_]
       :()
     ]
   ];
+ };
+
+/
+* @brief Publish a message applying a filter to a channel and a topic.
+* @param channel {symbol}: Channel to which publish a message.
+* @param topic {symbol}: Topic of the message.
+* @param message {any}: Message to send.
+* @param is_async {bool}: Flag to publish a message asynchronously.
+\
+.cmng_api.publish:{[channel;topic;table;message;is_async]
+  ($[is_async; neg; ::] raze CONSUMER_FILTERS[((channel; topic); (channel; `all))][`sockets]) @\: (`.cmng_api.update; table; (.z.p; topic; MY_ACCOUNT_NAME; message));
+ }
+
+/
+* @brief Start a private chat with a specific user.
+* @param name {symbol}: User name to connect to the connection manager.
+* @param target {symbol}: Traget user name to talk with.
+* @param is_requester {bool}: Indicates if the caller is a requester.
+* @param channel {symbol}: Chennel to use for the private chat.
+*  - null: Arbitrary value for a request.
+*  - other: Unique value assigned by the connection manager.
+\
+.cmng_api.start_private_chat:{[name;target;is_requester;channel]
+
+  // Create a chat table and update function if they do not exist
+  if[not `MESSAGE_BOX in tables[];
+    MESSAGE_BOX:: flip `time`topic`sender`message!"pss*"$\: ();
+    .cmng_api.update: {[table;message]
+      table insert (.z.p; message 1; message 2; message 3)
+    }
+  ];
+
+  // If this is a request, overwrite with a unique value generated on the connection manager.
+  if[is_requester;
+    result: CONNECTION_MANAGER_SOCKET (`.cmng.process_private_chat_request; name; target);
+    $[10h ~ type result;
+      // Propagate an error
+      'result;
+      // Overwrite the channel value
+      channel: last result
+    ]
+  ];
+  // Close the temporary handle to this process which was opened remotely.
+  if[.z.w > 0i; hclose .z.w];
+  // Register as both producer and consumer.
+  .cmng_api.register_as_consumer[name; channel; enlist `user_chat];
+  .cmng_api.register_as_producer[name; channel];
+  // Register the channel for the target.
+  PRIVATE_MESSAGE_CHANNEL[target]: channel;
+
+  if[is_requester;
+    // Notify the target to start chat
+    handle: hopen hsym `$":" sv -1 _ result;
+    neg[handle] (`.cmng_api.start_private_chat; target; name; 0b; channel)
+  ];
+ };
+
+/
+* @brief Publish a message to a specific user in private chat.
+* @param message {string}: Text message to send.
+\
+.cmng_api.publish_private: {[target;message]
+  .cmng_api.publish[PRIVATE_MESSAGE_CHANNEL target; `user_chat; `MESSAGE_BOX; enlist message; 1b]
  };
